gitdiff/requirements.txt
Django == 4.0.5
djangorestframework == 3.13.1
django-cors-headers == 3.14.0
python-gitlab == 3.15.0
celery==5.1.2
redis==3.5.3